Arm WIN live paper on the Genial feed and keep Enviar from sending on a real account.

+++ src/trader/domain.py
@@ -80,7 +80,7 @@ class Trade:
     result: str
     reason: str
     hour: int
-    contracts: int = 1
+    contracts: float = 1
 
     def to_dict(self) -> dict[str, Any]:
         return {
